% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{StatinReInitiation0829v2-package}
\alias{StatinReInitiation0829v2}
\alias{StatinReInitiation0829v2-package}
\title{StatinReInitiation0829v2: A Package Skeleton for Comparative Effectiveness Studies}
\description{
A skeleton package, to be used as a starting point when implementing comparative effect studies.
}
\keyword{internal}
